refactor(cookieprovider path change): cookieprovider path change
cookieprovider path change
